Addition of trajcluster to physplit.
Trajcluster is a module based on python-cluster
(see http://sourceforge.net/projects/python-cluster/). It uses a hierarchical
clustering algorithm (so far...) to cluster a set of trajectories based on one
of a number of measures. The measures implemented so far are:

1. Great Circle Distance: measure of the distance between points on
comparable trajectories, as per Jorba et.al. (2004) Does not take height into account.
2. Mahalanobis Distance: Uses comparison of three spatial variables in
cartesian coordinates, and their covariances, to define a scale-independent
measure of similarity.

Other functions supporting all of this have been included in this commit